requirements/base.txt
wheel==0.30.0
Django==1.11.6
django-environ==0.4.4
whitenoise==3.3.1
django-crispy-forms==1.6.1
django-model-utils==3.0.0
Pillow==4.2.1
argon2-cffi==16.3.0
psycopg2==2.7.3.1
awesome-slugify==1.6.5
pytz==2017.2
django-redis==4.8.0
redis>=2.10.5
celery==3.1.25
rcssmin==1.0.6
django-compressor==2.2
djangorestframework==3.6.4
Markdown==2.6.9
django-filter==1.0.4
djangorestframework-jwt==1.11.0
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
djangorestframework-queryfields==1.0.0
